+++ packages/specifications/src/projet/lauréat/garantiesFinancières/effacerHistorique/effacerHistoriqueGarantiesFinancières.feature
@@ -15,6 +15,7 @@ Fonctionnalité: Effacer tout l'historique de garanties financières d'un projet
         Et il ne devrait pas y avoir de garanties financières actuelles pour le projet "Du boulodrome de Marseille"
         Et un historique des garanties financières devrait être consultable pour le projet "Du boulodrome de Marseille"
 
+    @select
     Scénario: Impossible d'effacer l'historique des garanties financières s'il n'y a aucun historique de garanties financières sur le projet
         Quand un admin efface l'historique des garanties financières pour le projet "Du boulodrome de Marseille"
         Alors l'utilisateur devrait être informé que "Il n'y a aucunes garanties financières sur ce projet"
